Minor refactor changes
Added code to get the left and right position of the first machine setting. This needs to be assigned to a variable and passed in as extra when machine setting is being newly created. Will update things based on reel size changes. Might not be needed this way but eh..