Correctly added left/right position helper method
Added machine property (machineDirection) to be used with the updateLeftAndRightPositions method.

The method updates left/right positions when the reel size is selected/changed on AddMachineSetting activity.

This feature requires at least one MachineSetting to exist as it uses the left and right positions with the reel size used; as a base for all changes.